Wait for log cleanup similarly to how we wait for app dir cleanup.
