+++ SingularityExecutor/src/main/java/com/hubspot/singularity/executor/task/SingularityExecutorTaskCleanup.java
@@ -1,21 +1,29 @@
 package com.hubspot.singularity.executor.task;
 
 import java.io.IOException;
+import java.nio.file.DirectoryStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.PathMatcher;
 import java.nio.file.Paths;
+import java.util.ArrayDeque;
+import java.util.ArrayList;
+import java.util.Deque;
 import java.util.List;
 
 import org.slf4j.Logger;
 
 import com.google.common.collect.ImmutableList;
 import com.hubspot.singularity.executor.config.SingularityExecutorConfiguration;
+import com.hubspot.singularity.executor.config.SingularityExecutorLogrotateAdditionalFile;
 import com.hubspot.singularity.executor.utils.DockerUtils;
 import com.hubspot.singularity.runner.base.shared.ExceptionChainParser;
 import com.hubspot.singularity.runner.base.shared.SimpleProcessManager;
 import com.spotify.docker.client.exceptions.ContainerNotFoundException;
 import com.spotify.docker.client.exceptions.DockerException;
 import com.spotify.docker.client.messages.ContainerInfo;
+
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 
 public class SingularityExecutorTaskCleanup {
 
@@ -33,7 +41,7 @@ public class SingularityExecutorTaskCleanup {
     this.dockerUtils = dockerUtils;
   }
 
-  public TaskCleanupResult cleanup(boolean cleanupTaskAppDirectory, boolean isDocker) {
+  public TaskCleanupResult cleanup(boolean cleanupTaskAppDirectory, boolean cleanupLogs, boolean isDocker) {
     final Path taskDirectory = Paths.get(taskDefinition.getTaskDirectory());
 
     boolean dockerCleanSuccess = true;
@@ -71,7 +79,13 @@ public class SingularityExecutorTaskCleanup {
       return TaskCleanupResult.WAITING;
     }
 
+    if (!cleanupLogs) {
+      log.info("Not finishing cleanup because log files have not been cleaned yet");
+      return TaskCleanupResult.WAITING;
+    }
+
     boolean cleanupTaskAppDirectorySuccess = cleanupTaskAppDirectory();
+    checkForLogrotateAdditionalFilesToDelete(taskDefinition);
 
     log.info("Cleaned up logs ({}) and task app directory ({})", logTearDownSuccess, cleanupTaskAppDirectorySuccess);
 
@@ -123,7 +137,7 @@ public class SingularityExecutorTaskCleanup {
           "rm",
           "-rf",
           pathToDelete
-          );
+      );
 
       new SimpleProcessManager(log).runCommand(cmd);
 
@@ -133,5 +147,48 @@ public class SingularityExecutorTaskCleanup {
     }
 
     return false;
+  }
+
+  private void checkForLogrotateAdditionalFilesToDelete(SingularityExecutorTaskDefinition taskDefinition) {
+    configuration.getLogrotateAdditionalFiles()
+        .stream()
+        .filter(SingularityExecutorLogrotateAdditionalFile::isDeleteInExecutorCleanup)
+        .forEach(toDelete -> {
+          String glob = String.format("glob:%s/%s", taskDefinition.getTaskDirectoryPath().toAbsolutePath(), toDelete.getFilename());
+
+          log.debug("Trying to delete {} for task {} using glob {}...", toDelete.getFilename(), taskDefinition.getTaskId(), glob);
+
+          try {
+            List<Path> matches = findGlob(taskDefinition.getTaskDirectoryPath().toAbsolutePath(), taskDefinition.getTaskDirectoryPath().getFileSystem().getPathMatcher(glob));
+            for (Path match : matches) {
+              Files.delete(match);
+              log.debug("Deleted {}", match);
+            }
+          } catch (IOException e) {
+            log.error("Unable to list files while trying to delete for {}", toDelete);
+          }
+        });
+  }
+
+  @SuppressFBWarnings(value = "RCN_REDUNDANT_NULLCHECK_WOULD_HAVE_BEEN_A_NPE", justification = "https://github.com/spotbugs/spotbugs/issues/259")
+  private List<Path> findGlob(Path path, PathMatcher matcher) throws IOException {
+    Deque<Path> stack = new ArrayDeque<>();
+    List<Path> matched = new ArrayList<>();
+
+    stack.push(path);
+
+    while (!stack.isEmpty()) {
+      try (DirectoryStream<Path> stream = Files.newDirectoryStream(stack.pop())) {
+        for (Path entry : stream) {
+          if (Files.isDirectory(entry)) {
+            stack.push(entry);
+          } else if (matcher.matches(entry)) {
+            matched.add(entry);
+          }
+        }
+      }
+    }
+
+    return matched;
   }
 }
